Add DataService draft

+++ app/src/main/java/laiss/pokemon/android/web/Dtos.kt
@@ -27,8 +27,8 @@ data class PokemonStatsDto(val base_stat: Int, val stat: PokemonStatDto)
 data class PokemonDto(
     val name: String,
     val id: Int,
-    val height: Int,
-    val weight: Int,
+    val height: Double,
+    val weight: Double,
     val sprites: PokemonSpritesDto,
     val types: List<PokemonTypesDto>,
     val stats: List<PokemonStatsDto>
